requirements.txt
asttokens==3.0.0
black==25.1.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
identify==2.6.7
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isort==6.0.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.2
nvidia-nccl-cu12==2.25.1
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pip-tools==7.4.1
platformdirs==4.3.6
polars==1.22.0
pre_commit==4.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
skimpy==0.0.18
stack-data==0.6.3
threadpoolctl==3.5.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typeguard==4.4.1
types-pytz==2025.1.0.20250204
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.29
virtualenv==20.29.1
wcwidth==0.2.13
xgboost==2.1.4
